requirements.txt
python-dotenv==1.0.1
Faker==33.1.0
fastapi==0.115.6
uvicorn==0.34.0
pytest==8.3.4
opencv-python==4.11.0.86
numpy==2.2.3
httpx==0.28.1
